[REFACTOR] Reestructure project to facilitate tests
