+++ assets_app/lib/components/tree_view/tree_node.dart
@@ -1,0 +1,15 @@
+import 'package:assets_app/models/resources.dart';
+
+class TreeNode {
+  final Resource data;
+  TreeNode? parent;
+  final List<TreeNode> children;
+  int depth;
+
+  TreeNode({
+    required this.data,
+    required this.parent,
+    this.children = const [],
+    this.depth = 0,
+  });
+}
