+++ assets_app/lib/screens/assets/view.dart
@@ -1,7 +1,7 @@
 import 'package:assets_app/components/appbar.dart';
 import 'package:assets_app/components/buttons.dart';
 import 'package:assets_app/components/inputs.dart';
-import 'package:assets_app/components/lazy_tree_view.dart';
+import 'package:assets_app/components/tree_view/lazy_tree_view.dart';
 import 'package:assets_app/models/companies.dart';
 import 'package:assets_app/screens/assets/controller.dart';
 import 'package:flutter/material.dart';
